чуть изменил структуру папок в shared WIP
