+++ project/libs/shared/core/src/index.ts
@@ -1,16 +1,20 @@
-
 export { Entity } from './lib/base/entity';
 
-export { Comment } from './lib/comment/comment.interface'
-export { File } from './lib/file/file.interface'
-export { Like } from './lib/like/like.interface'
-export { TGenericPost } from './lib/post/post-pack.interface'
-export { TPostStatusList, PostStatus } from './lib/post/post-status.enum'
-export { PostContent, TPostContentList } from './lib/post/post-type.enum';
+export { User } from './lib/interfaces/user.interface';
+export { AuthUser } from './lib/interfaces/auth-user.interface';
+export { Post } from './lib/interfaces/post.interface';
+export { Comment } from './lib/interfaces/comment.interface';
+export { File } from './lib/interfaces/file.interface';
+export { Like } from './lib/interfaces/like.interface';
+export { TGenericPost } from './lib/types/post-pack.type';
+export { TPostStatusList } from './lib/types/post-status.type';
+export { PostStatus } from './lib/enums/post-status.enum';
+export { TPostContentList } from './lib/types/post-type.type';
+export { PostContent } from './lib/enums/post-type.enum';
 
-export { User } from './lib/user/user.interface';
-export { AuthUser } from './lib/user/auth-user.interface';
-export { StorableEntity } from './lib/repository/storable-entity.interface';
-export { EntityFactory } from './lib/repository/entity-factory.interface';
-export { SortDirection } from './lib/interfaces/sort-direction.interface';
+export { StorableEntity } from './lib/interfaces/storable-entity.interface';
+export { EntityFactory } from './lib/interfaces/entity-factory.interface';
+export { SortDirection } from './lib/enums/sort-direction.enum';
 export { PaginationResult } from './lib/interfaces/pagination.interface';
+export { Token } from './lib/interfaces/token.interface';
+export { TokenPayload } from './lib/interfaces/token-payload.interface';
