+++ project/libs/shared/core/src/lib/enums/post-status.enum.ts
@@ -2,5 +2,3 @@ export const PostStatus = {
   Posted: 'posted',
   Draft: 'draft',
 } as const;
-
-export type TPostStatusList = typeof PostStatus[keyof typeof PostStatus];
