+++ project/libs/shared/core/src/lib/enums/post-type.enum.ts
@@ -5,5 +5,3 @@ export const PostContent = {
   Quote: 'quote',
   Text: 'text'
 } as const;
-
-export type TPostContentList = typeof PostContent[keyof typeof PostContent];
