+++ project/libs/shared/core/src/lib/interfaces/auth-user.interface.ts
