+++ project/libs/shared/core/src/lib/interfaces/entity-factory.interface.ts
