+++ project/libs/shared/core/src/lib/interfaces/file.interface.ts
