+++ project/libs/shared/core/src/lib/interfaces/like.interface.ts
