+++ project/libs/shared/core/src/lib/interfaces/post-link.interface.ts
