+++ project/libs/shared/core/src/lib/interfaces/post-photo.interface.ts
