+++ project/libs/shared/core/src/lib/interfaces/post-text.interface.ts
