+++ project/libs/shared/core/src/lib/interfaces/storable-entity.interface.ts
