+++ project/libs/shared/core/src/lib/interfaces/token-payload.interface.ts
@@ -1,0 +1,5 @@
+export interface TokenPayload {
+  sub: string;
+  email: string;
+  login: string;
+}
