--- project/libs/shared/core/src/lib/repository/token-payload.interface.ts
@@ -1,9 +1,0 @@
-import { UserRole } from '../types/user-role.enum';
-
-export interface TokenPayload {
-  sub: string;
-  email: string;
-  role: UserRole;
-  lastname: string;
-  firstname: string;
-}
