+++ project/libs/shared/core/src/lib/types/post-status.type.ts
@@ -1,0 +1,3 @@
+import { PostStatus } from '../enums/post-status.enum';
+
+export type TPostStatusList = typeof PostStatus[keyof typeof PostStatus];
