+++ project/libs/shared/core/src/lib/types/post-type.type.ts
@@ -1,0 +1,3 @@
+import { PostContent } from '../enums/post-type.enum';
+
+export type TPostContentList = typeof PostContent[keyof typeof PostContent];
